Refactored parse_runner, and moved functions into a new file named parse_core.py
